+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # dependfix
 
+## [0.3.3](https://github.com/dependfix/dependfix/compare/dependfix@0.3.2...dependfix@0.3.3) (2026-08-25)
+
+### ⚙️ 依赖更新
+
+* bump `@dependfix/core` to 0.3.0 (was 0.2.1)
+* bump `@dependfix/engine` to 0.2.0 (was 0.1.3)
+
 ## [0.3.2](https://github.com/dependfix/dependfix/compare/dependfix@0.2.0...dependfix@0.3.2) (2026-08-12)